Project5 Update 12/1/24 Complete
Completed Project 5 which goals were to read in a prompt using an unordered_map (mainly speeches, or parts of books) and count the frequency of unique tokens ("The" and "the" are different, "man," and man." are different) and then sort the data in descending order using a multimap. Finally, the sorted map and the indices of the sorted map are outputted in correspondence to the input text. For example, if the sorted map had the word "the" as the most common word, meaning the index of the word "the" would be index 1, anytime the word "the" occurred in the text, the output would match the index of the word and output 1.